SYNC-1374 Be certain file is missing

+++ modules/util/sync-engine/src/com/liferay/sync/engine/util/FileUtil.java
@@ -182,6 +182,12 @@ public class FileUtil {
 			filePath.toString(), startTime);
 
 		for (SyncFile deletedSyncFile : deletedSyncFiles) {
+			if (!Files.notExists(
+					Paths.get(deletedSyncFile.getFilePathName()))) {
+
+				continue;
+			}
+
 			if (deletedSyncFile.getTypePK() == 0) {
 				SyncFileService.deleteSyncFile(deletedSyncFile, false);
